feat: replacing LS with IndexedDB in AuthProvider part 2 (completed)
